Feat14ПР

+++ controllers/cards.js
@@ -1,11 +1,12 @@
 const NotFoundError = require('../errors/NotFoundError');
-const NotValidCodeError = require('../errors/NotValidCodeError');
+const BadRequestError = require('../errors/BadRequestError');
 const Card = require('../models/card');
+const NoAccessError = require('../errors/NoAccessError');
 
 // сработает при GET-запросе на URL /cards
 module.exports.getCards = (_req, res, next) => {
   Card.find({})
-    .populate('owner')
+    // .populate('owner')
     .then((cards) => res.send({ cards }))
     .catch((err) => {
       next(err);
@@ -14,15 +15,13 @@ module.exports.getCards = (_req, res, next) => {
 
 // сработает при POST-запросе на URL /cards
 module.exports.createCard = (req, res, next) => {
-  const owner = req.user._id; // _id станет доступен
+  const owner = req.user._id;
   const { name, link } = req.body;
   Card.create({ name, link, owner })
     .then((card) => res.send({ card }))
     .catch((err) => {
-      // eslint-disable-next-line no-console
-      console.log(err.name);
       if (err.name === 'ValidationError') {
-        next(new NotValidCodeError('Переданы некорректные данные id'));
+        next(new BadRequestError('Переданы некорректные данные id'));
       } else {
         next(err);
       }
@@ -31,24 +30,21 @@ module.exports.createCard = (req, res, next) => {
 
 // сработает при DELETE-запросе на URL /cards/:cardId
 module.exports.deleteCard = (req, res, next) => {
-  // const owner = req.user._id; // _id станет доступен
+  const owner = req.user._id;
   const { cardId } = req.params;
   Card.findById(cardId).orFail(new NotFoundError('Карточка не найдена'))
-    // eslint-disable-next-line consistent-return
     .then((card) => {
       // Проверка на принадлежность карточки пользователю
-      // if (owner.toString() !== card.owner.toString()) {
-      //   return res.send({
-      //     message: `Пользователь с ID ${owner} не является владельцем данной карточки`,
-      //   });
-      // }
+      if (owner.toString() !== card.owner.toString()) {
+        next(new NoAccessError(`Пользователь с ID ${owner} не является владельцем данной карточки`));
+      }
       Card.findByIdAndRemove(cardId).then(() => {
         res.send({ message: `Карточка с ID ${card.id} удалена` });
       });
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        next(new NotValidCodeError('Передан несуществующий ID карточки'));
+        next(new BadRequestError('Передан несуществующий ID карточки'));
       } else if (err.name === 'ValidationError') {
         next(new NotFoundError('Карточка по указанному _id не найдена'));
       } else {
@@ -68,9 +64,9 @@ module.exports.likeCard = (req, res, next) => {
     .then((card) => res.send(card))
     .catch((err) => {
       if (err.name === 'CastError') {
-        next(new NotValidCodeError('Передан несуществующий ID карточки'));
+        next(new BadRequestError('Передан несуществующий ID карточки'));
       } else if (err.name === 'ValidationError') {
-        next(new NotValidCodeError(err.message));
+        next(new BadRequestError(err.message));
       } else {
         next(err);
       }
@@ -90,9 +86,9 @@ module.exports.dislikeCard = (req, res, next) => {
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        next(new NotValidCodeError('Передан несуществующий ID карточки'));
+        next(new BadRequestError('Передан несуществующий ID карточки'));
       } else if (err.name === 'ValidationError') {
-        next(new NotValidCodeError(err.message));
+        next(new BadRequestError(err.message));
       } else {
         next(err);
       }
